fix showing fd status.
when running a command, run() is spawned to run the command.  it
creates the fileio's used by the program.  run() sends its pid to
the caller when it will succeed.  but after that it creates the
fileio's, and the spawner was looking at the fileio's before that.
now run() syncs (sends its pid) a bit later.