pretty much a rewrite.
some code still survives.  the biggest changes:

- no split windows for now.  perhaps splitting stdin & stderr will
be implemented in the future.  but then it won't be like it was
(internally keep track of all i/o per command that was ever run),
only new i/o would be redirected to the newly split windows.

- no per-command i/o history anymore.  we still have command (the
text that was executed) history though.

- text typed is always put in the history, regardless of whether
it was executed as a command by the shell, or send to the stdin of
a command.

- many ^x shortcuts have gone.  many options too for that matter.
but wm/run has gained a wm/sh-like menu for toggling autoscroll &
displaying markup (colors).

various small fixes & improvements have also been added.  probably
some new bugs too unfortunately...