fix bug (showing up as tk error messages) when text is added while having the last text selected.
colors are drawn with text(9) tags.  when appending a string to the
text widget, tk gives the new text the same tags as the last
character.  to get the right tags, we deleted all tags and added
the new one.  but the "sel" tag, that indicates the selection,
cannot be removed.  so now we check for "sel", and for the common
case, we don't remove & add the tag if it the right tag is already
set.

thanks to jas for explaining the problem.